requirements.txt
# 網頁自動測試工具所需的依賴項目

# 網頁抓取
requests>=2.28.0
beautifulsoup4>=4.11.0

# Selenium 自動化測試
selenium>=4.8.0

# HTML 解析
lxml>=4.9.0

# 型別提示支援
typing-extensions>=4.4.0

# ChromeDriver 自動管理（推薦，解決瀏覽器顯示問題）
webdriver-manager>=3.8.0

# 環境變數配置
python-dotenv>=1.0.0 